Add state lock and only update if the memory has changed.
Adding a threading lock to make sure the state is not read while the component is updating. This can happen if it is present in another component ._component and updates are called recursively.
Added a check to only update if the length of the memory has changed, otherwise keep the state the same. This saves API calls.

PiperOrigin-RevId: 625615473
Change-Id: I41644ed9e3bf630521aafe5fbeb7367e7bd28d9f