Improve importance model.
The importance model now can be contextualized by retrieving other memories to provide scale for the decision of how to assign importance to each new memory.

This change also removes a parameter on the formative memories generator which overrode the importance value assigned by memory.add. If that behavior is desirable, a more consistent way to achieve it would be to use a Constant importance model.

PiperOrigin-RevId: 657156830
Change-Id: Icf5425fed4c3519d80c50ba6f4b10ede7df84865