Create v2 metrics and update the `riverbend_elections` colab example to use entity_agents and v2 components, including the newly updated metric components.
This also renames the two metrics to highlight that they are rather `context_free` at the moment. This has not changed how they work. It just highlights a way in which they should be improved in the future (they could incorporate more context, e.g. other components or memories). And this renames `goal_achievement` to `rationality`, but the question the metric asks does not change. It should have been named rationality all along since it asks about whether the agent took an action intended to move toward the goal, not whether the goal was achieved. This change also removes the OpinionOfOtherMetric from the elections colab for now. We will reimplement a new version of it in a different way later.

PiperOrigin-RevId: 661273623
Change-Id: Ib5c84e95f75d912dc87f2c7b0b5a10210f2a614f