Change map_parallel to return a Sequence instead of an Iterator
Previously, if there was an error it wouldn't be surfaced unless the returned iterator was consumed (the usual `ThreadPoolExecutor.map` behavior). This was probably surprising to people not used to `ThreadPoolExecutor.map` (and didn't match the docstring).

With this change, the any errors raised the workers will be surfaced at the `run_parallel` call site.  To distinguish it from `ThreadPoolExecutor.map`, I renamed it `run_parallel`.

PiperOrigin-RevId: 670187027
Change-Id: I277297c41b9e20b4131bd576214be3f44392e9d9